Move ui/field_formats to src/plugins/data (#44883)
