+++ src/legacy/core_plugins/kibana/common/field_formats/types/__tests__/bytes.js
@@ -19,7 +19,7 @@
 
 import expect from '@kbn/expect';
 import { createBytesFormat } from '../bytes';
-import { FieldFormat } from '../../../../../../ui/field_formats/field_format';
+import { FieldFormat } from '../../../../../../../plugins/data/common/field_formats';
 
 const BytesFormat = createBytesFormat(FieldFormat);
