+++ src/legacy/core_plugins/kibana/common/field_formats/types/__tests__/color.js
@@ -19,7 +19,7 @@
 
 import expect from '@kbn/expect';
 import { createColorFormat } from '../color';
-import { FieldFormat } from '../../../../../../ui/field_formats/field_format';
+import { FieldFormat } from '../../../../../../../plugins/data/common/field_formats';
 
 const ColorFormat = createColorFormat(FieldFormat);
