+++ src/legacy/ui/field_formats/index.ts
@@ -18,4 +18,3 @@
  */
 
 export { fieldFormatsMixin } from './mixin/field_formats_mixin';
-export { FieldFormat } from './field_format';
