+++ src/legacy/ui/public/courier/search_source/search_source.js
@@ -80,7 +80,7 @@ import { SearchRequestProvider } from '../fetch/request';
 import { searchRequestQueue } from '../search_request_queue';
 import { FetchSoonProvider } from '../fetch';
 import { FieldWildcardProvider } from '../../field_wildcard';
-import { getHighlightRequest } from '../../../../../plugins/data/common/highlight';
+import { getHighlightRequest } from '../../../../../plugins/data/common/field_formats';
 
 const FIELDS = [
   'type',
