+++ src/legacy/ui/public/field_editor/lib/__tests__/get_default_format.test.js
@@ -18,7 +18,7 @@
  */
 
 import { getDefaultFormat } from '../get_default_format';
-import { FieldFormat } from '../../../../field_formats';
+import { FieldFormat } from '../../../../../../plugins/data/common/field_formats';
 import { createNumberFormat } from '../../../../../core_plugins/kibana/common/field_formats/types/number';
 
 const Format = createNumberFormat(FieldFormat);
