+++ src/legacy/ui/public/registry/field_formats.js
@@ -19,7 +19,7 @@
 
 import _ from 'lodash';
 import chrome from '../chrome';
-import { FieldFormat } from '../../field_formats/field_format';
+import { FieldFormat } from '../../../../plugins/data/common/field_formats';
 import { IndexedArray } from '../indexed_array';
 
 class FieldFormatRegistry extends IndexedArray {
