+++ src/legacy/ui/public/visualize/loader/pipeline_helpers/utilities.ts
@@ -21,8 +21,9 @@ import { i18n } from '@kbn/i18n';
 import { identity } from 'lodash';
 import { AggConfig, Vis } from 'ui/vis';
 import { SerializedFieldFormat } from 'src/plugins/expressions/common/expressions/types/common';
-// @ts-ignore
-import { FieldFormat } from '../../../../field_formats/field_format';
+
+import { FieldFormat } from '../../../../../../plugins/data/common/field_formats';
+
 // @ts-ignore
 import { tabifyGetColumns } from '../../../agg_response/tabify/_get_columns';
 import chrome from '../../../chrome';
