+++ src/plugins/data/common/field_formats/utils/as_pretty_string.ts
