+++ src/plugins/data/common/field_formats/utils/highlight/highlight_tags.ts
