+++ src/plugins/data/common/index.ts
@@ -19,3 +19,4 @@
 
 export * from './timefilter/types';
 export * from './query/types';
+export * from './field_formats';
