+++ x-pack/legacy/plugins/reporting/export_types/csv/server/__tests__/execute_job.js
@@ -10,7 +10,7 @@ import sinon from 'sinon';
 import nodeCrypto from '@elastic/node-crypto';
 
 import { CancellationToken } from '../../../../common/cancellation_token';
-import { FieldFormat } from  '../../../../../../../../src/legacy/ui/field_formats/field_format';
+import { FieldFormat } from '../../../../../../../../src/plugins/data/common/field_formats';
 import { FieldFormatsService } from  '../../../../../../../../src/legacy/ui/field_formats/mixin/field_formats_service';
 import { createStringFormat } from  '../../../../../../../../src/legacy/core_plugins/kibana/common/field_formats/types/string';
